requirements.txt
anthropic>=0.40.0
streamlit>=1.40.0
gitpython>=3.1.40
plotly>=5.24.0
networkx>=3.3
pydantic>=2.9.0
python-dotenv>=1.0.0
